chore(pkg/auth): split auth code into multiple files
